+++ src/components/LayerTree/LayerTreeRenameRow.vue
@@ -2,6 +2,7 @@
 import { computed, useTemplateRef, watch } from 'vue'
 import { tv } from 'tailwind-variants'
 import { nodeIcon } from '@/app/editor/icons'
+import LayerTreeRowShell from './LayerTreeRowShell.vue'
 import LayerTreeDisclosure from './LayerTreeDisclosure.vue'
 import { useLayerTreeUI } from './ui'
 
@@ -30,12 +31,17 @@ watch(renameInput, (input) => {
 </script>
 
 <template>
-  <div
+  <LayerTreeRowShell
+    :pad-left="padLeft"
     data-slot="rename-row"
     :class="styles.renameRow({ class: ui?.renameRow })"
-    :style="{ paddingLeft: padLeft }"
   >
     <LayerTreeDisclosure
+      :label="node.name"
+      :ui="{
+        disclosure: styles.disclosure({ class: ui?.disclosure }),
+        placeholder: styles.disclosurePlaceholder({ class: ui?.disclosurePlaceholder })
+      }"
       :expanded="expanded"
       :visible="hasChildren"
       @toggle="actions.toggleExpand"
@@ -55,5 +61,5 @@ watch(renameInput, (input) => {
       @blur="renameControls.commit(node.id, $event)"
       @keydown.stop="renameControls.onKeydown"
     />
-  </div>
+  </LayerTreeRowShell>
 </template>